backport from baseq3a : Fix renderer animation at high level times (floatfix)

+++ omega_gamecode/code/cgame/cg_marks.c
@@ -290,8 +290,8 @@ void CG_AddMarks( void ) {
 typedef struct particle_s {
 	struct particle_s *next;
 
-	float time;
-	float endtime;
+	int time;
+	int endtime;
 
 	vec3_t org;
 	vec3_t vel;
@@ -312,7 +312,7 @@ typedef struct particle_s {
 	float start;
 	float end;
 
-	float startfade;
+	int startfade;
 	qboolean rotate;
 	int snum;
